tests: added stub for ipam cache in watcher package + refactoring around it
