+++ calico-vpp-agent/tests/mocks/calico/ippools.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mocks
+package calico
 
 import (
 	"context"
